Continue refactoring towards date processing
